added.plot {flexplot}	R Documentation

Create an added variable plot

Description

Create an added variable plot

Usage


added.plot(formula, data, method = "loess", ...)

Arguments

formula
A formula with exactly two predictors and an outcome variable

data
The dataset used

method
The smoothing method. Defaults to "loess"

...
Other parameters passed to flexplot

Details

This function first residualizes the outcome variable based on the first variables listed in the formula. The mean of the outcome variable is then added to the residuals (to maintain the interpretation of the variable), then the function plots the residuals against the second variable.

Author(s)

Dustin Fife

See Also

flexplot

Examples


data(exercise_data)
added.plot(weight.loss~motivation + therapy.type, data=exercise_data)

[Package flexplot version 0.8.8 Index]