avengers {flexplot}	R Documentation

Simulated Statistics on the Final Avengers Battle

Description

A dataset containing combat attributes of almost 812 fighters in the final Avengers battle

Usage


avengers

Format

A data frame with 812 rows and 13 variables:

iq

Intelligence

agility

weighted scores on an obstacle course

speed

Speed in running the 40 meter dash

strength

Pounds lifted in a benchpress

damage.resistance

Amount of skin deflection (in mm) when slapped with a frozen fish

flexibility

Number of inches past their toes they can reach

willpower

Length of time they wait at a DMV for a driver's license

ptsd

Score on a Likert-scale PTSD questionnaire

north_south

whether the individual was assigned to fight on the north or south battlefield

died

Whether the person died at the final battle

kills

Number of enemies killed in the final battle

injuries

Number of injuries sustained. Anything more than 5 is marked as a 5

minutes.fighting

Number of minutes spent fighting before dying or giving up

shots.taken

Number of shots (with a gun) or punches thrown at an enemy

superpower

Whether the individual has a superpower

[Package flexplot version 0.8.9 Index]