birthweight {flexplot}	R Documentation

Dataset seeks to examine what can affect birthweight.

Description

This dataset examines how various variables, especially related to parents, can affect the birthweight of a baby.

Usage


birthweight

Format

A data frame with xx rows and 16 variables:

id

observation id

headcirumference

Circumference of baby's head.

length

description

Birthweight

Weight at time of baby's head.

Gestation

Length of time baby was gestated.

smoker

is the mother a smoker (1=yes)

motherage

Age of mother.

mnocig

number of cigarettes smoked by mother

mheight

mother's height

mppwt

mother pre pregnancy weight

fage

father's age

fedyrs

father's years of education

fnocig

father number of cigarettes

fheight

father height

lowbwt

binary indicator of whether it's considered low weight or not

mage35

mother's age, centered on 35

LowBirthWeight

Same as above, but labeled

[Package flexplot version 0.8.11 Index]