criminal_data {flexplot}	R Documentation

Criminal dataset

Description

This is a simulated dataset that looks at how various variables affect convictions and rape instances

Usage


criminal_data

Format

A data frame with xx rows and 6 variables:

ses

Socioeconomic status of the subject

empathy

description

depression

Whether the subject is depressed

rape

description

convictions

How many times the subject has been convicted of a crime

aggression

How aggressive the subject is

[Package flexplot version 0.8.3.2 Index]