exercise_data {flexplot}	R Documentation

Exercise dataset

Description

Simulated data that sees how weight loss varies as a function of various predictors, and treatment condition

Usage


exercise_data

Format

A data frame with xx rows and 13 variables:

therapy.type

description

health

description

motivation

description

weight.loss

description

gender

description

satisfaction

description

rewards

description

income

description

muscle.gain

description

muscle.gain.missing

description

six.mo.weight

description

[Package flexplot version 0.8.8 Index]