glinmod {flexplot}	R Documentation

Perform a general linear model

Description

Perform a general linear model

Usage


glinmod(f, data, plot = TRUE, ...)

Arguments

f
An equation of the form y~x1 + x2 + x3, etc.

data
The dataset containing the variables the user wishes to model

plot
Should results be plotted?

...
Other parameters passed to visualize (which is then passed to flexplot)

Details

A general linear model (GLM) is a statistical model of the form Y = BX, where X can be any number of numeric or categorical predictors. The GLM subsumes the most common statistical models, including t-tests, ANOVAs, regressions, etc. The function will automatically detect whether the variables are numeric or categorical and generate estimates and graphics accordingly.

Value

An object containing the information from estimates

Author(s)

Dustin Fife

See Also

estimates, visualize, flexplot, model.comparison

Examples


## load the exercise dataset
data(exercise_data)
data = exercise_data
glinmod(weight.loss~gender, data=exercise_data)
## pass parameters to flexplot
glinmod(weight.loss~gender + motivation, data=exercise_data, 
	se=FALSE, method="lm")

[Package flexplot version 0.7.3 Index]