icc {flexplot}	R Documentation

Compute the ICC from a lmer (package lme4) model

Description

This function will extract the variances from a mixed model and output the value of the ICC

Usage


icc(model)

Arguments

model
a lmer model

Details

Recall that the equation for the ICC is var(school)/[var(school)+ var(person)]. This forumla extracts the necessary variances from the mixed model and computes it.

Value

the ICC

Author(s)

Dustin Fife

Examples


data(alcuse)
require(lme4)
model = lmer(ALCUSE~1|ID, data=alcuse)
icc(model)

[Package flexplot version 0.8.11 Index]