make.formula {flexplot}	R Documentation

Convert strings to a formula

Description

make.formula is a function that easily converts a set of strings into a formula. It requires two arguments: a single response variable, and a vector of strings. See examples.

Usage


make.formula(response, predictors, random = NULL)

Arguments

response
a single string used on the left side of a formula

predictors
a string (or a vector of strings) representing the predictors. Each will be separated by a plus sign.

random
a string that indicates the random component in an lmer-like object (e.g., "(1|group)"). Defaults to NULL.

Value

a formula object

Author(s)

Dustin Fife

Examples


k = data.frame(matrix(rnorm(100), ncol=5))
names(k) = LETTERS[1:5]
formula = make.formula("A", LETTERS[2:5])
formula
lm(formula, data=k)
#do a random model
make.formula("A", LETTERS[2:5], random="(1|group)")

[Package flexplot version 0.8.8 Index]