mixed.mod.visual {flexplot}	R Documentation

Display a graphic of a mixed model

Description

Display a graphic of a mixed model

Usage


mixed.mod.visual(formula, data, model, n = 6, jitter = F)

Arguments

formula
A flexplot friendly formula, though it can only include one predictor

data
The dataset

model
The model

n
The number of IDs to sample and put in different panels

jitter
Should the data be jittered?

Details

This function samples the ID variable and plots the bivariate relationship, depending on the mixed model chosen

Value

A plot

Author(s)

Dustin Fife

[Package flexplot version 0.8.11 Index]