paranormal {flexplot}	R Documentation

Simulated Dataset About Experiences with the Paranormal

Description

A dataset containing combat attributes of almost 812 fighters in the final Avengers battle. This dataset illustrates several types of problems one might encounter for univariate variables, including bimodality, zero-inflated data, outliers, mixed up labels, etc.

Usage


paranormal

Format

A data frame with 1000 rows and 7 variables:

conviction

Degree of conviction they have about the existence of the paranormal

fear

How much they fear being kidnapped by aliens

time

How much time they spend a year researching the paranormal

kidnapped

Whether they've been kidnapped by aliens

experiences.type

What type of experiences they have had with the paranormal? Can be "eerie feeling," "presence watching", "saw UFO", "saw ghost", or "heard voice"

income

How much money they make

age

Age of respondent

gender

Gender of respondent

political

Political Ideology of respondent

[Package flexplot version 0.8.8 Index]