rsq_change {flexplot}	R Documentation

Compare the improvement in R squared from a baseline model

Description

Compare the improvement in R squared from a baseline model

Usage


rsq_change(full, reduced)

Arguments

full
the full mixed model

reduced
the reduced mixed model (typically a random effect ANOVA model)

Details

Compare the improvement in fit from a baseline model, using mixed models

Value

A vector containing the proportion change in fit for each random effects parameter

Author(s)

Dustin Fife

Examples


data(math)
require(lme4)
baseline.model = lmer(MathAch~1 + (1|School), data=math)
full.model = lmer(MathAch~SES + (SES|School), data=math)
rsq_change(full.model, baseline.model)

[Package flexplot version 0.8.9 Index]