sensitivity.table {flexplot}	R Documentation

Compute sensitivity/specificity/etc.

Description

Compute sensitivity/specificity/etc.

Usage


sensitivity.table(object)

Arguments

object
an object that can be predicted (e.g., glm). Note the thing to be predicted must have only two outcomes

Details

This function computes sensitivity, specificity, positive/negative predictive value and accuracy and reports them as a list.

Author(s)

Dustin Fife

[Package flexplot version 0.8.3.2 Index]