uni.plot {flexplot}	R Documentation

Plot univariate distribution of a variable

Description

A function that automatically plots either a barchart or a histogram (depending on data type)

Usage


uni.plot(variable, d = NULL, numeric = NULL)

Arguments

variable
A string indicating the name of the variable to plot

d
The name of the dataset (which contains the variable the user wishes to plot)

numeric
Logical. Is the variable of interest numeric? (Meaning, should a histogram be plotted?)

Details

Ideally, the user will provide the data type for this function (by saying numeric = T, or numeric = F) Omitting that, the function will decide for the user. If the variable is coded as numeric AND there are more than five unique values of the variable, it will generate a histogram. If the variable is a factor OR it has five or less unique values, it will output a barchart.

Value

A plot

Author(s)

Dustin Fife

Examples


distress = sample(1:10, size=22, replace=TRUE)
uni.plot(distress, numeric=TRUE, d=NULL)

[Package flexplot version 0.7.3 Index]